+++ config/config_test.go
@@ -24,6 +24,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"github.com/uber/jaeger-lib/metrics"
 
 	"github.com/uber/jaeger-client-go"
 )
@@ -63,7 +64,7 @@ func TestNewSamplerProbabilistic(t *testing.T) {
 
 func TestDefaultSampler(t *testing.T) {
 	cfg := &SamplerConfig{MaxOperations: 10}
-	s, err := cfg.NewSampler("x", jaeger.NewMetrics(jaeger.NullStatsReporter, nil))
+	s, err := cfg.NewSampler("x", jaeger.NewNullMetrics())
 	require.NoError(t, err)
 	rcs, ok := s.(*jaeger.RemotelyControlledSampler)
 	require.True(t, ok, "converted to RemotelyControlledSampler")
@@ -72,10 +73,10 @@ func TestDefaultSampler(t *testing.T) {
 
 func TestDefaultConfig(t *testing.T) {
 	cfg := Configuration{}
-	_, _, err := cfg.New("", jaeger.NullStatsReporter)
+	_, _, err := cfg.New("", Metrics(metrics.NullFactory), Logger(jaeger.NullLogger))
 	require.EqualError(t, err, "no service name provided")
 
-	_, closer, err := cfg.New("testService", jaeger.NullStatsReporter)
+	_, closer, err := cfg.New("testService")
 	defer closer.Close()
 	require.NoError(t, err)
 }