+++ metrics_test.go
@@ -23,23 +23,34 @@ package jaeger
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/uber/jaeger-lib/metrics"
+	"github.com/uber/jaeger-lib/metrics/testutils"
 )
 
 func TestNewMetrics(t *testing.T) {
 	tags := map[string]string{"lib": "jaeger"}
 
-	m := &Metrics{}
-	err := initMetrics(m, nil, tags)
-	require.NoError(t, err)
+	factory := metrics.NewLocalFactory(0)
+	m := NewMetrics(factory, tags)
 
-	m = NewMetrics(nil, tags)
 	require.NotNil(t, m.SpansSampled, "counter not initialized")
 	require.NotNil(t, m.ReporterQueueLength, "gauge not initialized")
-	require.NotEmpty(t, m.SpansSampled.tags)
-	assert.Equal(t, "jaeger.spans", m.SpansSampled.name)
-	assert.Equal(t, "sampling", m.SpansSampled.tags["group"])
-	assert.Equal(t, "y", m.SpansSampled.tags["sampled"])
-	assert.Equal(t, "jaeger", m.SpansSampled.tags["lib"])
+
+	m.SpansSampled.Inc(1)
+	m.ReporterQueueLength.Update(11)
+	testutils.AssertCounterMetrics(t, factory,
+		testutils.ExpectedMetric{
+			Name:  "jaeger.spans",
+			Tags:  map[string]string{"group": "sampling", "lib": "jaeger", "sampled": "y"},
+			Value: 1,
+		},
+	)
+	testutils.AssertGaugeMetrics(t, factory,
+		testutils.ExpectedMetric{
+			Name:  "jaeger.reporter-queue",
+			Tags:  map[string]string{"lib": "jaeger"},
+			Value: 11,
+		},
+	)
 }