+++ sampler_test.go
@@ -27,6 +27,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/uber/jaeger-lib/metrics"
+	mTestutils "github.com/uber/jaeger-lib/metrics/testutils"
 
 	"github.com/uber/jaeger-client-go/testutils"
 	"github.com/uber/jaeger-client-go/thrift-gen/sampling"
@@ -265,12 +267,12 @@ func TestAdaptiveSamplerUpdate(t *testing.T) {
 	assert.Len(t, sampler.samplers, 2)
 }
 
-func initAgent(t *testing.T) (*testutils.MockAgent, *RemotelyControlledSampler, *InMemoryStatsCollector) {
+func initAgent(t *testing.T) (*testutils.MockAgent, *RemotelyControlledSampler, *metrics.LocalFactory) {
 	agent, err := testutils.StartMockAgent()
 	require.NoError(t, err)
 
-	stats := NewInMemoryStatsCollector()
-	metrics := NewMetrics(stats, nil)
+	metricsFactory := metrics.NewLocalFactory(0)
+	metrics := NewMetrics(metricsFactory, nil)
 
 	initialSampler, _ := NewProbabilisticSampler(0.001)
 	sampler := NewRemotelyControlledSampler(
@@ -284,11 +286,11 @@ func initAgent(t *testing.T) (*testutils.MockAgent, *RemotelyControlledSampler, 
 	)
 	sampler.Close() // stop timer-based updates, we want to call them manually
 
-	return agent, sampler, stats
+	return agent, sampler, metricsFactory
 }
 
 func TestRemotelyControlledSampler(t *testing.T) {
-	agent, sampler, stats := initAgent(t)
+	agent, sampler, metricsFactory := initAgent(t)
 	defer agent.Close()
 
 	initSampler, ok := sampler.sampler.(*ProbabilisticSampler)
@@ -300,7 +302,13 @@ func TestRemotelyControlledSampler(t *testing.T) {
 			SamplingRate: 1.5, // bad value
 		}})
 	sampler.updateSampler()
-	assert.EqualValues(t, 1, stats.GetCounterValue("jaeger.sampler", "phase", "parsing", "state", "failure"))
+	mTestutils.AssertCounterMetrics(t, metricsFactory,
+		mTestutils.ExpectedMetric{
+			Name:  "jaeger.sampler",
+			Tags:  map[string]string{"state": "failure", "phase": "parsing"},
+			Value: 1,
+		},
+	)
 	_, ok = sampler.sampler.(*ProbabilisticSampler)
 	assert.True(t, ok)
 	assert.Equal(t, initSampler, sampler.sampler, "Sampler should not have been updated")
@@ -311,12 +319,11 @@ func TestRemotelyControlledSampler(t *testing.T) {
 			SamplingRate: testDefaultSamplingProbability, // good value
 		}})
 	sampler.updateSampler()
-	assertMetrics(t, stats, []expectedMetric{
-		{[]string{"jaeger.sampler", "phase", "parsing", "state", "failure"}, 1},
-		{[]string{"jaeger.sampler", "state", "retrieved"}, 1},
-		{[]string{"jaeger.sampler", "state", "updated"}, 1},
-	})
-
+	mTestutils.AssertCounterMetrics(t, metricsFactory, []mTestutils.ExpectedMetric{
+		{Name: "jaeger.sampler", Tags: map[string]string{"state": "failure", "phase": "parsing"}, Value: 1},
+		{Name: "jaeger.sampler", Tags: map[string]string{"state": "retrieved"}, Value: 1},
+		{Name: "jaeger.sampler", Tags: map[string]string{"state": "updated"}, Value: 1},
+	}...)
 	_, ok = sampler.sampler.(*ProbabilisticSampler)
 	assert.True(t, ok)
 	assert.NotEqual(t, initSampler, sampler.sampler, "Sampler should have been updated")
@@ -351,7 +358,7 @@ func TestRemotelyControlledSampler(t *testing.T) {
 }
 
 func TestUpdateSampler(t *testing.T) {
-	agent, sampler, stats := initAgent(t)
+	agent, sampler, metricsFactory := initAgent(t)
 	defer agent.Close()
 
 	initSampler, ok := sampler.sampler.(*ProbabilisticSampler)
@@ -360,21 +367,21 @@ func TestUpdateSampler(t *testing.T) {
 	tests := []struct {
 		probabilities      map[string]float64
 		defaultProbability float64
-		statTags           []string
-		statCount          int64
+		statTags           string
+		statCount          int
 		isErr              bool
 	}{
 		{
 			map[string]float64{testOperationName: 1.1},
 			testDefaultSamplingProbability,
-			[]string{"phase", "parsing", "state", "failure"},
+			"|phase=parsing|state=failure",
 			1,
 			true,
 		},
 		{
 			map[string]float64{testOperationName: testDefaultSamplingProbability},
 			testDefaultSamplingProbability,
-			[]string{"state", "updated"},
+			"|state=updated",
 			1,
 			false,
 		},
@@ -384,35 +391,35 @@ func TestUpdateSampler(t *testing.T) {
 				testFirstTimeOperationName: testDefaultSamplingProbability,
 			},
 			testDefaultSamplingProbability,
-			[]string{"state", "updated"},
+			"|state=updated",
 			2,
 			false,
 		},
 		{
 			map[string]float64{testOperationName: 1.1},
 			testDefaultSamplingProbability,
-			[]string{"phase", "updating", "state", "failure"},
+			"|phase=updating|state=failure",
 			1,
 			true,
 		},
 		{
 			map[string]float64{"new op": 1.1},
 			testDefaultSamplingProbability,
-			[]string{"phase", "updating", "state", "failure"},
+			"|phase=updating|state=failure",
 			2,
 			true,
 		},
 		{
 			map[string]float64{testOperationName: testDefaultSamplingProbability},
 			0.5,
-			[]string{"state", "updated"},
+			"|state=updated",
 			3,
 			false,
 		},
 		{
 			map[string]float64{testOperationName: testDefaultSamplingProbability},
 			1.1,
-			[]string{"phase", "updating", "state", "failure"},
+			"|phase=updating|state=failure",
 			3,
 			true,
 		},
@@ -440,7 +447,11 @@ func TestUpdateSampler(t *testing.T) {
 		agent.AddSamplingStrategy("client app", res)
 		sampler.updateSampler()
 
-		assert.EqualValues(t, test.statCount, stats.GetCounterValue("jaeger.sampler", test.statTags...))
+		mTestutils.AssertCounterMetrics(t, metricsFactory,
+			mTestutils.ExpectedMetric{
+				Name: "jaeger.sampler" + test.statTags, Value: test.statCount,
+			},
+		)
 		if test.isErr {
 			continue
 		}
@@ -480,7 +491,7 @@ func TestMaxOperations(t *testing.T) {
 }
 
 func TestSamplerQueryError(t *testing.T) {
-	agent, sampler, stats := initAgent(t)
+	agent, sampler, metricsFactory := initAgent(t)
 	defer agent.Close()
 
 	// override the actual handler
@@ -493,7 +504,14 @@ func TestSamplerQueryError(t *testing.T) {
 
 	sampler.updateSampler()
 	assert.Equal(t, initSampler, sampler.sampler, "Sampler should not have been updated due to query error")
-	assert.EqualValues(t, 1, stats.GetCounterValue("jaeger.sampler", "phase", "query", "state", "failure"))
+
+	mTestutils.AssertCounterMetrics(t, metricsFactory,
+		mTestutils.ExpectedMetric{
+			Name:  "jaeger.sampler",
+			Tags:  map[string]string{"phase": "query", "state": "failure"},
+			Value: 1,
+		},
+	)
 }
 
 type fakeSamplingManager struct{}