+++ tracer_test.go
@@ -29,22 +29,24 @@ import (
 	"github.com/opentracing/opentracing-go/ext"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
+	"github.com/uber/jaeger-lib/metrics"
+	"github.com/uber/jaeger-lib/metrics/testutils"
 
 	"github.com/uber/jaeger-client-go/utils"
 )
 
 type tracerSuite struct {
 	suite.Suite
-	tracer opentracing.Tracer
-	closer io.Closer
-	stats  *InMemoryStatsCollector
+	tracer         opentracing.Tracer
+	closer         io.Closer
+	metricsFactory *metrics.LocalFactory
 }
 
 var IP uint32 = 1<<24 | 2<<16 | 3<<8 | 4
 
 func (s *tracerSuite) SetupTest() {
-	s.stats = NewInMemoryStatsCollector()
-	metrics := NewMetrics(s.stats, nil)
+	s.metricsFactory = metrics.NewLocalFactory(0)
+	metrics := NewMetrics(s.metricsFactory, nil)
 
 	s.tracer, s.closer = NewTracer("DOOP", // respect the classics, man!
 		NewConstSampler(true),
@@ -66,6 +68,7 @@ func TestTracerSuite(t *testing.T) {
 }
 
 func (s *tracerSuite) TestBeginRootSpan() {
+	s.metricsFactory.Clear()
 	startTime := time.Now()
 	s.tracer.(*tracer).timeNow = func() time.Time { return startTime }
 	someID := uint64(12345)
@@ -89,12 +92,12 @@ func (s *tracerSuite) TestBeginRootSpan() {
 	sp.Finish()
 	s.NotNil(ss.duration)
 
-	assertMetrics(s.T(), s.stats, []expectedMetric{
-		{[]string{"jaeger.spans", "group", "lifecycle", "state", "started"}, 1},
-		{[]string{"jaeger.spans", "group", "lifecycle", "state", "finished"}, 1},
-		{[]string{"jaeger.spans", "group", "sampling", "sampled", "y"}, 1},
-		{[]string{"jaeger.traces", "sampled", "y", "state", "started"}, 1},
-	})
+	testutils.AssertCounterMetrics(s.T(), s.metricsFactory, []testutils.ExpectedMetric{
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "started"}, Value: 1},
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "finished"}, Value: 1},
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "sampling", "sampled": "y"}, Value: 1},
+		{Name: "jaeger.traces", Tags: map[string]string{"sampled": "y", "state": "started"}, Value: 1},
+	}...)
 }
 
 func (s *tracerSuite) TestStartRootSpanWithOptions() {
@@ -107,18 +110,19 @@ func (s *tracerSuite) TestStartRootSpanWithOptions() {
 }
 
 func (s *tracerSuite) TestStartChildSpan() {
+	s.metricsFactory.Clear()
 	sp1 := s.tracer.StartSpan("get_address")
 	sp2 := s.tracer.StartSpan("get_street", opentracing.ChildOf(sp1.Context()))
 	s.Equal(sp1.(*span).context.spanID, sp2.(*span).context.parentID)
 	sp2.Finish()
 	s.NotNil(sp2.(*span).duration)
 	sp1.Finish()
-	assertMetrics(s.T(), s.stats, []expectedMetric{
-		{[]string{"jaeger.spans", "group", "sampling", "sampled", "y"}, 2},
-		{[]string{"jaeger.traces", "sampled", "y", "state", "started"}, 1},
-		{[]string{"jaeger.spans", "group", "lifecycle", "state", "started"}, 2},
-		{[]string{"jaeger.spans", "group", "lifecycle", "state", "finished"}, 2},
-	})
+	testutils.AssertCounterMetrics(s.T(), s.metricsFactory, []testutils.ExpectedMetric{
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "sampling", "sampled": "y"}, Value: 2},
+		{Name: "jaeger.traces", Tags: map[string]string{"sampled": "y", "state": "started"}, Value: 1},
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "started"}, Value: 2},
+		{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "finished"}, Value: 2},
+	}...)
 }
 
 func (s *tracerSuite) TestTraceStartedOrJoinedMetrics() {
@@ -130,7 +134,7 @@ func (s *tracerSuite) TestTraceStartedOrJoinedMetrics() {
 		{false, "n"},
 	}
 	for _, test := range tests {
-		s.stats.Clear()
+		s.metricsFactory.Clear()
 		s.tracer.(*tracer).sampler = NewConstSampler(test.sampled)
 		sp1 := s.tracer.StartSpan("parent", ext.RPCServerOption(nil))
 		sp2 := s.tracer.StartSpan("child1", opentracing.ChildOf(sp1.Context()))
@@ -143,13 +147,13 @@ func (s *tracerSuite) TestTraceStartedOrJoinedMetrics() {
 		s.Equal(test.sampled, sp1.Context().(SpanContext).IsSampled())
 		s.Equal(test.sampled, sp2.Context().(SpanContext).IsSampled())
 
-		assertMetrics(s.T(), s.stats, []expectedMetric{
-			{[]string{"jaeger.spans", "group", "sampling", "sampled", test.label}, 3},
-			{[]string{"jaeger.spans", "group", "lifecycle", "state", "started"}, 3},
-			{[]string{"jaeger.spans", "group", "lifecycle", "state", "finished"}, 3},
-			{[]string{"jaeger.traces", "state", "started", "sampled", test.label}, 1},
-			{[]string{"jaeger.traces", "state", "joined", "sampled", test.label}, 1},
-		})
+		testutils.AssertCounterMetrics(s.T(), s.metricsFactory, []testutils.ExpectedMetric{
+			{Name: "jaeger.spans", Tags: map[string]string{"group": "sampling", "sampled": test.label}, Value: 3},
+			{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "started"}, Value: 3},
+			{Name: "jaeger.spans", Tags: map[string]string{"group": "lifecycle", "state": "finished"}, Value: 3},
+			{Name: "jaeger.traces", Tags: map[string]string{"sampled": test.label, "state": "started"}, Value: 1},
+			{Name: "jaeger.traces", Tags: map[string]string{"sampled": test.label, "state": "joined"}, Value: 1},
+		}...)
 	}
 }